remove llama unittests into a new file
